fix: deregister spam on render and on match bug
The fix is done as follows:
Set isConnecting to null on initial render so that no deregister is sent on initial render.

Pass number of retries between UI and Matching-service via sockets. -> Failsafe for infinite loop

Remove deregister on socket "matchFound" event -> this was a inclusion that was not deleted once matching service was updated. On match, matching-service was handle deletion of waitees in the database on it's own. The above contributed to deregister spam once a match is found.